+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGBlaze.java
@@ -19,7 +19,6 @@ import com.craftcosta.jailrules.rpgcraftcosta.entities.CustomEntityType;
 import com.craftcosta.jailrules.rpgcraftcosta.entities.RPGMob;
 import com.craftcosta.jailrules.rpgcraftcosta.entities.utils.AttackType;
 import com.craftcosta.jailrules.rpgcraftcosta.entities.utils.MobBehaviour;
-import com.craftcosta.jailrules.rpgcraftcosta.entities.utils.ProjectileType;
 
 /**
  *
@@ -46,7 +45,7 @@ public class RPGBlaze extends RPGMob{
      * @param exp
      * @param ap
      */
-    public RPGBlaze(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap) {
+    public RPGBlaze(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap) {
         this.level = level;
         this.name = name;
         this.type = type;
