+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGChicken.java
@@ -47,7 +47,7 @@ public class RPGChicken extends RPGMob{
      * @param ap
      * @param baby
      */
-    public RPGChicken(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap, boolean baby) {
+    public RPGChicken(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap, boolean baby) {
         this.level = level;
         this.name = name;
         this.type = type;
