+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGEnderDragon.java
@@ -46,7 +46,7 @@ public class RPGEnderDragon extends RPGMob{
      * @param exp
      * @param ap
      */
-    public RPGEnderDragon(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap) {
+    public RPGEnderDragon(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap) {
         this.level = level;
         this.name = name;
         this.type = type;
