+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGGuardian.java
@@ -48,7 +48,7 @@ public class RPGGuardian extends RPGMob{
      * @param ap
      * @param gType
      */
-    public RPGGuardian(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap, GuardianType gType) {
+    public RPGGuardian(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap, GuardianType gType) {
         this.level = level;
         this.name = name;
         this.type = type;
