+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGHorse.java
@@ -53,7 +53,7 @@ public class RPGHorse extends RPGMob{
      * @param htype
      * @param hVariant
      */
-    public RPGHorse(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap,boolean baby, HorseType htype, HorseVariant hVariant) {
+    public RPGHorse(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap,boolean baby, HorseType htype, HorseVariant hVariant) {
         this.level = level;
         this.name = name;
         this.type = type;
