+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGRabbit.java
@@ -51,7 +51,7 @@ public class RPGRabbit extends RPGMob{
      * @param rtype
      * @param baby
      */
-    public RPGRabbit(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap,RabbitType rtype,boolean baby) {
+    public RPGRabbit(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap,RabbitType rtype,boolean baby) {
         this.level = level;
         this.name = name;
         this.type = type;
