+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGSheep.java
@@ -50,7 +50,7 @@ public class RPGSheep extends RPGMob{
      * @param scolor
      * @param baby
      */
-    public RPGSheep(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap,SheepColor scolor, boolean baby) {
+    public RPGSheep(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap,SheepColor scolor, boolean baby) {
         this.level = level;
         this.name = name;
         this.type = type;
