+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGSilverfish.java
@@ -45,7 +45,7 @@ public class RPGSilverfish extends RPGMob{
      * @param exp
      * @param ap
      */
-    public RPGSilverfish(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap) {
+    public RPGSilverfish(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap) {
         this.level = level;
         this.name = name;
         this.type = type;
