+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGSkeleton.java
@@ -49,7 +49,7 @@ public class RPGSkeleton extends RPGMob{
      * @param ap
      * @param stype
      */
-    public RPGSkeleton(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap,SkeletonType stype) {
+    public RPGSkeleton(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap,SkeletonType stype) {
         this.level = level;
         this.name = name;
         this.type = type;
