+++ src/com/craftcosta/jailrules/rpgcraftcosta/entities/rpgentities/RPGVillager.java
@@ -51,7 +51,7 @@ public class RPGVillager extends RPGMob{
      * @param vtype
      * @param baby
      */
-    public RPGVillager(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, double money, double exp, double ap,VillagerType vtype,boolean baby) {
+    public RPGVillager(int level, String name, CustomEntityType type, AttackType aType, MobBehaviour mType, double damageattack, double movementspeed, double knockback, double followrange, double maxhealth, double attackspeed, double rangeddamage, float rangedstrength, long money, long exp, double ap,VillagerType vtype,boolean baby) {
         this.level = level;
         this.name = name;
         this.type = type;
